# docs/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/credit-cards/{id}": {
      "patch": {
        "operationId": "CreditCardsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/UpdateCreditCardDto" } }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreditCardResponseDto_Output" }
              }
            }
          }
        },
        "tags": ["CreditCards"]
      },
      "delete": {
        "operationId": "CreditCardsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "format": "uuid", "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "tags": ["CreditCards"]
      }
    },
    "/v1/credit-cards/{id}/invoices": {
      "get": {
        "operationId": "CreditCardsController_invoice",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "description": "Mês da fatura no formato YYYY-MM.",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-(0[1-9]|1[0-2])$",
              "example": "2026-07"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/InvoiceResponseDto_Output" }
              }
            }
          }
        },
        "tags": ["CreditCards"]
      }
    },
...
    "/v1/insights/by-category": {
      "get": {
        "operationId": "InsightsController_byCategory",
        "parameters": [
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Limite superior exclusivo do intervalo [from, to).",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "description": "Limite inferior inclusivo do intervalo [from, to).",
            "schema": { "format": "date-time", "type": "string" }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/CategoryInsightDto_Output" }
                }
              }
            }
          }
        },
        "tags": ["Insights"]
      }
    },
    "/v1/insights/balance-series": {
      "get": {
        "operationId": "InsightsController_balanceSeries",
        "parameters": [
          {
            "name": "to",
            "required": true,
            "in": "query",
            "description": "Limite superior exclusivo do intervalo [from, to).",
            "schema": { "format": "date-time", "type": "string" }
          },
          {
            "name": "from",
            "required": true,
            "in": "query",
            "description": "Limite inferior inclusivo do intervalo [from, to).",
            "schema": { "format": "date-time", "type": "string" }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/BalancePointDto_Output" }
                }
              }
            }
          }
        },
        "tags": ["Insights"]
      }
    },
...
  },
  "info": {
    "title": "Fluxo API",
    "description": "API financeira autoritativa para os clientes Fluxo.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
...
      "CreditCardResponseDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "accountId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": { "type": "string" },
          "limitCents": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 },
          "closingDay": { "type": "integer", "minimum": 1, "maximum": 31 },
          "dueDay": { "type": "integer", "minimum": 1, "maximum": 31 },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
          }
        },
        "required": [
          "id",
          "userId",
          "accountId",
          "name",
          "limitCents",
          "closingDay",
          "dueDay",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "UpdateCreditCardDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": { "type": "string", "minLength": 1, "maxLength": 80 },
          "limitCents": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 },
          "closingDay": { "type": "integer", "minimum": 1, "maximum": 31 },
          "dueDay": { "type": "integer", "minimum": 1, "maximum": 31 }
        },
        "minProperties": 1,
        "additionalProperties": false
      },
      "InvoiceResponseDto_Output": {
        "type": "object",
        "properties": {
          "creditCardId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "month": { "type": "string", "pattern": "^\\d{4}-(0[1-9]|1[0-2])$" },
          "totalCents": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 },
          "status": { "type": "string", "enum": ["OPEN", "CLOSED"] }
        },
        "required": ["creditCardId", "month", "totalCents", "status"],
        "additionalProperties": false
      },
...
      "CategoryInsightDto_Output": {
        "type": "object",
        "properties": {
          "categoryId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              { "type": "null" }
            ]
          },
          "categoryName": { "anyOf": [{ "type": "string" }, { "type": "null" }] },
          "totalCents": { "type": "integer", "minimum": 0, "maximum": 9007199254740991 }
        },
        "required": ["categoryId", "categoryName", "totalCents"],
        "additionalProperties": false
      },
      "BalancePointDto_Output": {
        "type": "object",
        "properties": {
          "day": { "type": "string", "pattern": "^\\d{4}-\\d{2}-\\d{2}$" },
          "cumulativeCents": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": ["day", "cumulativeCents"],
        "additionalProperties": false
      },
...
    }
  }
}